Change authentication requirements
In the past, we have required folks to be a member of a specific team in a
specific org. Now, we just require folks to have their GitHub account linked
to their Microsoft corp-net alias. That means in practice that most Microsoft
folks will have access to our planning.

Hopefully, this is simpler than having to maintain a dedicated team just for
this site.