Allow cross-linking between GitHub and Azure DevOps
This removes the ability to configure queries in the subscription
configuration. The expectation is that those are by having GitHub issues
referencing Azure DevOps queries or individual items directly.

We want to be able to index individual work items, not just queries. Hence, we
need to model the notions that Azure DevOps work items can exist without
queries.

This also removes the grouping of work items under a fake node that represents
the query. We'll replace that with the ability to link to Azure DevOps items
and queries from GitHub markdown.